chapter13/python-scripts/requirements.txt
neo4j
openai>=1.6.1
python-dotenv>=1.0.0